Corección de Bugs Del Mapa y agregar icono de la aplicación ademas de cambiarle el nombre a Fisio App
