+++ app/src/main/AndroidManifest.xml
@@ -8,9 +8,10 @@
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
-        android:icon="@mipmap/ic_launcher"
+        android:icon="@drawable/logoyidam"
         android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
+        android:roundIcon="@drawable/logoyidam"
+
         android:supportsRtl="true"
         android:theme="@style/Theme.ProyectoTitulacion">
 
@@ -20,7 +21,7 @@
 
         <!-- Pantalla principal -->
         <activity
-            android:name=".LoginActivity"
+            android:name=".Login.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -29,22 +30,22 @@
         </activity>
 
         <activity
-            android:name=".RegisterActivity"
+            android:name=".Registro.RegisterActivity"
             android:exported="false" />
         <activity
-            android:name=".PerfilActivity"
+            android:name=".PerfilUsuario.PerfilActivity"
             android:exported="false" />
 
         <activity
-            android:name=".HomeActivity"
+            android:name=".Mapa.HomeActivity"
             android:exported="false" />
         <activity
-            android:name=".ChatActivity"
+            android:name=".ChatBot.ChatActivity"
             android:exported="false" />
 
         <!-- AÑADE ESTA DECLARACIÓN PARA MainActivity -->
         <activity
-            android:name=".MainActivity"
+            android:name=".Citas.MainActivity"
             android:exported="false" />
         <!-- FIN DE LA DECLARACIÓN AÑADIDA -->
