+++ app/src/main/java/com/example/proyectotitulacion/Citas/HelperCitas.java
@@ -1,4 +1,4 @@
-package com.example.proyectotitulacion;
+package com.example.proyectotitulacion.Citas;
 
 import android.content.ContentValues;
 import android.content.Context;
