+++ app/src/main/res/layout/activity_chat.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background_pantallasgenerales"
-    tools:context=".ChatActivity">
+    tools:context=".ChatBot.ChatActivity">
 
     <!-- RecyclerView para mostrar los mensajes del chat -->
     <androidx.recyclerview.widget.RecyclerView
