+++ app/src/main/res/layout/activity_home.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background_pantallasgenerales"
-    tools:context=".HomeActivity">
+    tools:context=".Mapa.HomeActivity">
 
     <androidx.core.widget.NestedScrollView
         android:id="@+id/scrollView"
