+++ app/src/main/res/layout/activity_perfil.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background_pantallasgenerales"
-    tools:context=".PerfilActivity">
+    tools:context=".PerfilUsuario.PerfilActivity">
 
     <TextView
         android:id="@+id/tvPerfilTitulo"
