+++ app/src/main/res/layout/login_principal_main.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg_gradient"
-    tools:context=".MainActivity">
+    tools:context=".Citas.MainActivity">
 
     <ImageView
         android:id="@+id/logoImage"
