+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">ProyectoTitulacion</string>
+    <string name="app_name">Fisio App</string>
     <string name="rama_main_hint">Home Rama Main</string>
     <string name="logo_description">Logo</string>
     <string name="username_hint">Usuario</string>
